backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
python-dotenv
pydantic
playwright
numpy
Pillow
supabase
googlesearch-python
beautifulsoup4
requests
typing-extensions
google-genai
